Rename database traits [skip ci]
