+++ tests/Mysql/AssignmentsStorageTest.php
@@ -6,5 +6,5 @@ namespace Yiisoft\Rbac\Db\Tests\Mysql;
 
 class AssignmentsStorageTest extends \Yiisoft\Rbac\Db\Tests\Base\AssignmentsStorageTest
 {
-    use MysqlTrait;
+    use DatabaseTrait;
 }
