+++ tests/Mysql/DatabaseTrait.php
@@ -2,20 +2,20 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Rbac\Db\Tests\Pgsql;
+namespace Yiisoft\Rbac\Db\Tests\Mysql;
 
 use Yiisoft\Cache\ArrayCache;
 use Yiisoft\Db\Cache\SchemaCache;
 use Yiisoft\Db\Connection\ConnectionInterface;
-use Yiisoft\Db\Pgsql\Connection;
-use Yiisoft\Db\Pgsql\Driver;
+use Yiisoft\Db\Mysql\Connection;
+use Yiisoft\Db\Mysql\Driver;
 
-trait PgsqlTrait
+trait DatabaseTrait
 {
     protected function makeDatabase(): ConnectionInterface
     {
-        $pdoDriver = new Driver('pgsql:host=127.0.0.1;dbname=yiitest;port=5432', 'root', 'root');
-        $pdoDriver->charset('UTF8');
+        $pdoDriver = new Driver('mysql:host=127.0.0.1;dbname=yiitest;port=3306', 'root');
+        $pdoDriver->charset('UTF8MB4');
 
         return new Connection($pdoDriver, new SchemaCache(new ArrayCache()));
     }
