+++ tests/Mysql/DbSchemaManagerTest.php
@@ -6,5 +6,5 @@ namespace Yiisoft\Rbac\Db\Tests\Mysql;
 
 class DbSchemaManagerTest extends \Yiisoft\Rbac\Db\Tests\Base\DbSchemaManagerTest
 {
-    use MysqlTrait;
+    use DatabaseTrait;
 }
