+++ tests/Oracle/DatabaseTrait.php
@@ -11,7 +11,7 @@ use Yiisoft\Db\Connection\ConnectionInterface;
 use Yiisoft\Db\Oracle\Connection;
 use Yiisoft\Db\Oracle\Driver;
 
-trait OracleTrait
+trait DatabaseTrait
 {
     protected function makeDatabase(): ConnectionInterface
     {
