+++ tests/Oracle/DbSchemaManagerTest.php
@@ -6,5 +6,5 @@ namespace Yiisoft\Rbac\Db\Tests\Oracle;
 
 class DbSchemaManagerTest extends \Yiisoft\Rbac\Db\Tests\Base\DbSchemaManagerTest
 {
-    use OracleTrait;
+    use DatabaseTrait;
 }
