+++ tests/Oracle/ItemsStorageTest.php
@@ -6,5 +6,5 @@ namespace Yiisoft\Rbac\Db\Tests\Oracle;
 
 class ItemsStorageTest extends \Yiisoft\Rbac\Db\Tests\Base\ItemsStorageTest
 {
-    use OracleTrait;
+    use DatabaseTrait;
 }
