+++ tests/Pgsql/DbSchemaManagerTest.php
@@ -6,5 +6,5 @@ namespace Yiisoft\Rbac\Db\Tests\Pgsql;
 
 class DbSchemaManagerTest extends \Yiisoft\Rbac\Db\Tests\Base\DbSchemaManagerTest
 {
-    use PgsqlTrait;
+    use DatabaseTrait;
 }
